Move user attribute registry from plugins
